refactor: SharedStatusType 위치 변경
