+++ backend/src/test/java/hanglog/expense/fixture/TripExpenseFixture.java
@@ -6,7 +6,7 @@ import static hanglog.expense.fixture.ExpenseFixture.USD_100_ACCOMMODATION_EXPEN
 import static hanglog.integration.IntegrationFixture.MEMBER;
 
 import hanglog.community.domain.type.PublishedStatusType;
-import hanglog.share.domain.type.SharedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Item;
 import hanglog.trip.domain.Trip;
