+++ backend/src/test/java/hanglog/trip/fixture/DayLogFixture.java
@@ -3,7 +3,7 @@ package hanglog.trip.fixture;
 import static hanglog.integration.IntegrationFixture.MEMBER;
 
 import hanglog.community.domain.type.PublishedStatusType;
-import hanglog.share.domain.type.SharedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Item;
 import hanglog.trip.domain.Trip;
