+++ backend/src/test/java/hanglog/trip/service/TripServiceTest.java
@@ -20,7 +20,7 @@ import hanglog.community.domain.PublishedTrip;
 import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.global.exception.BadRequestException;
 import hanglog.member.domain.repository.MemberRepository;
-import hanglog.share.domain.type.SharedStatusType;
+import hanglog.trip.domain.type.SharedStatusType;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.repository.CustomDayLogRepository;
